+++ temba/utils/tests.py
@@ -22,6 +22,8 @@ from mock import patch, PropertyMock
 from openpyxl import load_workbook
 from temba.contacts.models import Contact, ContactField, ContactGroup, ContactGroupCount, ExportContactsTask
 from temba.locations.models import AdminBoundary
+from temba.msgs.models import Msg, SystemLabel
+from temba.flows.models import FlowRun
 from temba.orgs.models import Org
 from temba.tests import TembaTest
 from temba_expressions.evaluator import EvaluationContext, DateStyle
@@ -41,9 +43,26 @@ from .profiler import time_monitor
 from .queues import start_task, complete_task, push_task, HIGH_PRIORITY, LOW_PRIORITY, nonoverlapping_task
 from .timezones import TimeZoneFormField, timezone_to_country_code
 from .voicexml import VoiceXMLException
+from . import decode_base64
 
 
 class InitTest(TembaTest):
+
+    def test_decode_base64(self):
+
+        self.assertEqual('This test\nhas a newline', decode_base64('This test\nhas a newline'))
+
+        self.assertEqual('Please vote NO on the confirmation of Gorsuch.',
+                         decode_base64('Please vote NO on the confirmation of Gorsuch.'))
+
+        self.assertEqual('Bannon Explains The World ...\n\u201cThe Camp of the Saints',
+                         decode_base64('QmFubm9uIEV4cGxhaW5zIFRoZSBXb3JsZCAuLi4K4oCcVGhlIENhbXAgb2YgdGhlIFNhaW50c+KA\r'))
+
+        self.assertEqual('the sweat, the tears and the sacrifice of working America',
+                         decode_base64('dGhlIHN3ZWF0LCB0aGUgdGVhcnMgYW5kIHRoZSBzYWNyaWZpY2Ugb2Ygd29ya2luZyBBbWVyaWNh\r'))
+
+        self.assertIn('I find them to be friendly',
+                      decode_base64('Tm93IGlzDQp0aGUgdGltZQ0KZm9yIGFsbCBnb29kDQpwZW9wbGUgdG8NCnJlc2lzdC4NCg0KSG93IGFib3V0IGhhaWt1cz8NCkkgZmluZCB0aGVtIHRvIGJlIGZyaWVuZGx5Lg0KcmVmcmlnZXJhdG9yDQoNCjAxMjM0NTY3ODkNCiFAIyQlXiYqKCkgW117fS09Xys7JzoiLC4vPD4/fFx+YA0KQUJDREVGR0hJSktMTU5PUFFSU1RVVldYWVphYmNkZWZnaGlqa2xtbm9wcXJzdHV2d3h5eg=='))
 
     def test_datetime_to_ms(self):
         d1 = datetime.datetime(2014, 1, 2, 3, 4, 5, tzinfo=pytz.utc)
@@ -1450,29 +1469,39 @@ class MakeTestDBTest(SimpleTestCase):
     allow_database_queries = True
 
     def tearDown(self):
+        Msg.objects.all().delete()
+        FlowRun.objects.all().delete()
+        SystemLabel.objects.all().delete()
         Org.objects.all().delete()
         User.objects.all().delete()
         Group.objects.all().delete()
         AdminBoundary.objects.all().delete()
 
     def test_command(self):
-        call_command('make_test_db', num_orgs=2, num_contacts=12, seed=123456)
+        call_command('make_test_db', num_orgs=3, num_contacts=30, num_runs=20, seed=1234)
 
-        org_1, org_2 = list(Org.objects.order_by('id'))
+        org1, org2, org3 = tuple(Org.objects.order_by('id'))
 
-        self.assertEqual(User.objects.count(), 10)  # 4 for each org + superuser + anonymous
-        self.assertEqual(ContactField.objects.count(), 12)  # 6 per org
-        self.assertEqual(ContactGroup.user_groups.count(), 20)  # 10 per org
-        self.assertEqual(Contact.objects.filter(is_test=True).count(), 8)  # 1 for each user
-        self.assertEqual(Contact.objects.filter(is_test=False).count(), 4)
+        def assertOrgCounts(qs, counts):
+            self.assertEqual([qs.filter(org=o).count() for o in (org1, org2, org3)], counts)
 
-        org_1_all_contacts = ContactGroup.system_groups.get(org=org_1, name="All Contacts")
+        self.assertEqual(User.objects.count(), 15)  # 4 for each org + superuser + anonymous + flow user
+        assertOrgCounts(ContactField.objects.all(), [6, 6, 6])
+        assertOrgCounts(ContactGroup.user_groups.all(), [10, 10, 10])
+        assertOrgCounts(Contact.objects.filter(is_test=True), [4, 4, 4])  # 1 for each user
+        assertOrgCounts(Contact.objects.filter(is_test=False), [17, 7, 6])
+        assertOrgCounts(FlowRun.objects.filter(contact__is_test=True), [12, 12, 12])  # each input template per org
+        assertOrgCounts(FlowRun.objects.filter(contact__is_test=False), [10, 4, 6])
+        assertOrgCounts(Msg.objects.filter(contact__is_test=False), [12, 4, 8])
 
-        self.assertEqual(org_1_all_contacts.contacts.count(), 3)
-        self.assertEqual(list(ContactGroupCount.objects.filter(group=org_1_all_contacts).values_list('count')), [(3,)])
+        org_1_all_contacts = ContactGroup.system_groups.get(org=org1, name="All Contacts")
+
+        self.assertEqual(org_1_all_contacts.contacts.count(), 17)
+        self.assertEqual(list(ContactGroupCount.objects.filter(group=org_1_all_contacts).values_list('count')), [(17,)])
+        self.assertEqual(SystemLabel.get_counts(org1), {'I': 0, 'W': 1, 'A': 0, 'O': 0, 'S': 11, 'X': 0, 'E': 0, 'C': 0})
 
         # same seed should generate objects with same UUIDs
-        self.assertEqual(ContactGroup.user_groups.order_by('id').first().uuid, 'cec602da-1406-e378-df14-b8d4a99b7cc4')
+        self.assertEqual(ContactGroup.user_groups.order_by('id').first().uuid, 'ea60312b-25f5-47a0-cac7-4fe0c2064f3e')
 
         # check can't be run again on a now non-empty database
         with self.assertRaises(CommandError):